Refactor: Move HealthComponent to separate file and add combine function
The HealthComponent has been moved from MiscComponents.kt to its own file, HealthComponent.kt.

A new function, `combineScoreComponents`, has been added to the ScoreComponent. This function allows combining two ScoreComponent instances by summing their scores.

A similar function, `combineHealthComponents`, has been added to the HealthComponent. This function allows combining two HealthComponent instances by summing their current health and maximum health values.

Unit tests for the `combineHealthComponents` function have been updated to reflect the changes.
